+++ src/doomstat.h
@@ -446,7 +446,7 @@ extern boolean hide_weapon;
 // [FG] centered weapon sprite
 extern int center_weapon;
 
-// [Nugget] /--------------------------
+// [Nugget] /-----------------------------------------------------------------
 
 extern boolean fauxdemo;
 extern boolean casual_play;
@@ -454,8 +454,9 @@ extern boolean casual_play;
 #define    CASUALPLAY(x) (casual_play ? (x) : 0)
 #define NOTCASUALPLAY(x) (casual_play ? (x) : 1)
 
-// General ------------------
+// General ----------------------------
 
+extern int gammacycle; // CFG-Only
 extern int wipe_type;
 extern int over_under;
 extern int jump_crouch;
@@ -482,6 +483,7 @@ enum {
 }; extern int chasecam_mode;
 extern int chasecam_distance;
 extern int chasecam_height;
+extern int chasecam_crosshair; // CFG-Only
 
 extern int menu_background_all;
 extern int no_menu_tint;
@@ -501,10 +503,11 @@ extern int a11y_weapon_flash;
 extern int a11y_weapon_pspr;
 extern int a11y_invul_colormap;
 
-// Weapons ------------------
+// Weapons ----------------------------
 
 extern int no_hor_autoaim;
 extern int switch_on_pickup;
+extern int always_bob; // CFG-Only
 extern int weapon_bobbing_percentage;
 
 enum {
@@ -518,15 +521,18 @@ enum {
 }; extern int bobbing_style;
 
 extern int weapon_inertia;
+extern int weapon_inertia_scale_pct; // CFG-Only
 extern int weaponsquat;
 extern int translucent_pspr;
 extern int show_berserk;
+extern int sx_fix; // CFG-Only
 
-// Status Bar/HUD -----------
+// Status Bar/HUD ---------------------
 
+extern int show_ssg; // CFG-Only
+extern int alt_arms;
+extern int blink_keys; // CFG-Only
 extern int smarttotals;
-extern int hudcolor_ms_incomp;
-extern int hudcolor_ms_comp;
 
 typedef enum {
   TIMER_USE,
@@ -537,34 +543,19 @@ typedef enum {
 } eventtimer_t;
 extern int event_timers[];
 
-// Enemies ------------------
+extern int hudcolor_time_scale;
+extern int hudcolor_total_time;
+extern int hudcolor_time;
+extern int hudcolor_event_timer;
+extern int hudcolor_kills;
+extern int hudcolor_items;
+extern int hudcolor_secrets;
+extern int hudcolor_ms_incomp;
+extern int hudcolor_ms_comp;
 
-extern int extra_gibbing_on;
-extern int bloodier_gibbing;
-extern int zdoom_item_drops;
+// Enemies ----------------------------
 
-// Messages -----------------
-
-extern int announce_milestones;
-
-// Key Bindings -------------
-
-extern int zoom_fov;
-extern int fancy_teleport;
-
-// CFG only -----------------
-
-extern int gammacycle;
-extern int menu_background_darkening;
-extern int chasecam_crosshair;
-extern int always_bob;
-extern int weapon_inertia_scale_pct;
-extern int sx_fix;
-extern int blink_keys;
-extern int show_ssg;
-extern int alt_arms;
-extern int automap_overlay_darkening;
-
+extern int extra_gibbing_on; // CFG-Only
 typedef enum {
   EXGIB_FIST,
   EXGIB_CSAW,
@@ -574,7 +565,19 @@ typedef enum {
 } extragibbing_t;
 extern int extra_gibbing[];
 
-extern int sp_chat;
+extern int bloodier_gibbing;
+extern int zdoom_item_drops;
+
+// Messages ---------------------------
+
+extern int announce_milestones;
+
+// Key Bindings -----------------------
+
+extern int zoom_fov;
+extern int fancy_teleport;
+
+// Miscellaneous (CFG-Only) -----------
 
 enum {
   SHOTPAL_NONE,
@@ -583,7 +586,11 @@ enum {
   SHOTPAL_BOTH,
 }; extern int screenshot_palette;
 
-// Doom Compatibility -------
+extern int menu_background_darkening;
+extern int automap_overlay_darkening;
+extern int sp_chat;
+
+// Doom Compatibility -----------------
 
 extern int comp_bruistarget;
 extern int comp_nomeleesnap;
@@ -605,7 +612,7 @@ extern int comp_deadoof;
 extern int comp_unusedpals;
 extern int comp_keypal;
 
-// [Nugget] --------------------------/
+// [Nugget] -----------------------------------------------------------------/
 
 // Doom-style printf